lab3_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]
passlib[bcrypt]
httpx
psycopg2-binary
neo4j
redis
elasticsearch==8.18.0
pymongo
python-multipart>=0.0.5